Convert plugins window tests
