+++ trview.app.tests/trview.app.tests.vcxproj
@@ -98,7 +98,6 @@
     <ClCompile Include="Windows\ConsoleManagerTests.cpp" />
     <ClCompile Include="Windows\LightsWindowManagerTests.cpp" />
     <ClCompile Include="Windows\LogWindowManagerTests.cpp" />
-    <ClCompile Include="Windows\PluginsWindowTests.cpp" />
     <ClCompile Include="Windows\RouteWindowManagerTests.cpp" />
     <ClCompile Include="Windows\ViewerTests.cpp" />
   </ItemGroup>
