+++ trview.app.tests/trview.app.tests.vcxproj.filters
@@ -173,9 +173,6 @@
     <ClCompile Include="PluginsWindowManagerTests.cpp">
       <Filter>Windows</Filter>
     </ClCompile>
-    <ClCompile Include="Windows\PluginsWindowTests.cpp">
-      <Filter>Windows</Filter>
-    </ClCompile>
     <ClCompile Include="Lua\Elements\Lua_StaticMeshTests.cpp">
       <Filter>Lua\Elements</Filter>
     </ClCompile>
