+++ trview.app.ui.tests/PluginsWindowTests.h
@@ -1,0 +1,5 @@
+#pragma once
+
+struct ImGuiTestEngine;
+
+void register_plugins_window_tests(ImGuiTestEngine* engine);
