+++ trview.app.ui.tests/trview.app.ui.tests.vcxproj.filters
@@ -87,6 +87,9 @@
     <ClCompile Include="LightsWindowTests.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="PluginsWindowTests.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\external\shared\imgui_app.h">
@@ -138,6 +141,9 @@
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="LightsWindowTests.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="PluginsWindowTests.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
